+++ Unpublished/VersionText/changelog.md
@@ -1,5 +1,12 @@
 ﻿# Version History of Version Text Mod
 
+## Version 4.2020.5.12
+
+* All fields made Volatile for API thread-safety.
+* Sheepy.VerText no longer change config when parameter is null.
+* Switched DebugConsoleAdder to use Harmony AccessTools.
+* Protect UIStateMainMenu_EnterState.Refresh with try catch.
+
 ## Version 4
 
 * Compatible with DebugConsole